+++ app/helpers/dorsale/link_helper.rb
@@ -5,5 +5,40 @@ module Dorsale
       link_to(name, options, html_options, &block)
     end
 
+    def web_link(text, opts = {})
+      return if text.to_s.blank?
+
+      href = text
+      href = "http://#{text}" unless text.include?("://")
+
+      link_to(text, href, opts)
+    end
+
+    def email_link(text, opts = {})
+      return if text.to_s.blank?
+
+      href = "mailto:#{text}"
+
+      link_to(text, href, opts)
+    end
+
+    def tel_link(text, opts = {})
+      return if text.to_s.blank?
+
+      href = "tel:#{text}"
+
+      link_to(text, href, opts)
+    end
+
+    def twitter_link(text, opts = {})
+      return if text.to_s.blank?
+
+      href = text
+      href = "https://twitter.com/#{text}" unless text.include?("twitter.com")
+      href = "https://#{text}" unless href.include?("://")
+
+      link_to(text, href, opts)
+    end
+
   end
 end